Move eigen_matrix_compare from util to common
