+++ drake/examples/Acrobot/test/urdfDynamicsTest.cpp
@@ -1,13 +1,12 @@
 #include "drake/common/drake_path.h"
 #include "drake/examples/Acrobot/Acrobot.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "gtest/gtest.h"
 
 using drake::RigidBodySystem;
 using drake::getRandomVector;
 using drake::GetDrakePath;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace examples {
