+++ drake/examples/Cars/test/simple_car_test.cc
@@ -9,10 +9,9 @@
 #include "drake/systems/cascade_system.h"
 #include "drake/systems/simulation_options.h"
 #include "drake/systems/vector.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/examples/Cars/system1_cars_vectors.h"
 
-using drake::util::MatrixCompareType;
 using drake::NullVector;
 using drake::SimulationOptions;
