+++ drake/examples/Pendulum/pendulum_run_swing_up.cc
@@ -15,11 +15,11 @@
 #include "drake/systems/plants/robot_state_tap.h"
 #include "drake/systems/Simulation.h"
 #include "drake/util/drakeAppUtil.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 
 using drake::SimpleLqrTrajectoryController;
 using drake::solvers::SolutionResult;
-using drake::util::MatrixCompareType;
+using drake::MatrixCompareType;
 
 typedef PiecewisePolynomial<double> PiecewisePolynomialType;
