+++ drake/examples/Pendulum/test/urdfDynamicsTest.cpp
@@ -1,14 +1,13 @@
 #include "drake/common/drake_path.h"
 #include "drake/examples/Pendulum/Pendulum.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 #include "gtest/gtest.h"
 
 using drake::GetDrakePath;
 using drake::getRandomVector;
 using drake::RigidBodySystem;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace examples {
