+++ drake/examples/Quadrotor/test/urdfDynamicsTest.cpp
@@ -4,13 +4,12 @@
 
 #include "drake/common/drake_path.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 
 using drake::GetDrakePath;
 using drake::getRandomVector;
 using drake::RigidBodySystem;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace examples {
