+++ drake/examples/kuka_iiwa_arm/test/kuka_iiwa_pd_control_test.cc
@@ -10,7 +10,7 @@
 #include "drake/systems/pd_control_system.h"
 #include "drake/systems/plants/BotVisualizer.h"
 #include "drake/systems/plants/robot_state_tap.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 
 using drake::AffineSystem;
 using drake::BotVisualizer;
@@ -19,7 +19,6 @@ using drake::PDControlSystem;
 using drake::RigidBodySystem;
 using drake::RobotStateTap;
 using Eigen::VectorXd;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace examples {
