+++ drake/math/test/autodiff_test.cc
@@ -7,15 +7,13 @@
 
 #include "drake/common/eigen_types.h"
 #include "drake/math/autodiff_gradient.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 
 using Eigen::MatrixXd;
 using Eigen::VectorXd;
 
 namespace drake {
 
-using util::CompareMatrices;
-using util::MatrixCompareType;
 
 namespace math {
 namespace {
