+++ drake/math/test/expmap_test.cc
@@ -5,7 +5,7 @@
 
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 
 using drake::math::initializeAutoDiff;
 using Eigen::AutoDiffScalar;
@@ -14,8 +14,6 @@ using Eigen::Vector4d;
 
 namespace drake {
 
-using util::CompareMatrices;
-using util::MatrixCompareType;
 
 namespace math {
 namespace {
