+++ drake/solvers/test/mosek_test.cc
@@ -9,11 +9,10 @@
 #include "drake/util/testUtil.h"
 #include "drake/solvers/mathematical_program.h"
 #include "drake/solvers/optimization.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/solvers/constraint.h"
 
 
-using drake::util::MatrixCompareType;
 namespace drake {
 namespace solvers {
 namespace {
