+++ drake/solvers/test/optimization_problem_test.cc
@@ -8,7 +8,7 @@
 #include "drake/solvers/optimization.h"
 #include "drake/solvers/snopt_solver.h"
 #include "drake/common/polynomial.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 #include "gtest/gtest.h"
 
@@ -25,7 +25,6 @@ using Eigen::VectorXd;
 
 using drake::solvers::detail::VecIn;
 using drake::solvers::detail::VecOut;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace solvers {
