+++ drake/systems/controllers/InstantaneousQPController.cpp
@@ -11,7 +11,7 @@
 #include "drake/solvers/fast_qp.h"
 #include "drake/systems/controllers/controlUtil.h"
 #include "drake/systems/plants/parser_urdf.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/lcmUtil.h"
 #include "drake/util/testUtil.h"
 #include "drake/util/yaml/yamlUtil.h"
@@ -23,7 +23,6 @@ const bool CHECK_CENTROIDAL_MOMENTUM_RATE_MATCHES_TOTAL_WRENCH = false;
 const bool PUBLISH_ZMP_COM_OBSERVER_STATE = true;
 
 using namespace Eigen;
-using drake::util::MatrixCompareType;
 
 #define LEG_INTEGRATOR_DEACTIVATION_MARGIN 0.07
 
@@ -563,8 +562,9 @@ void checkCentroidalMomentumMatchesTotalWrench(
       world_momentum_matrix * qdd + world_momentum_matrix_dot_times_v;
 
   std::string explanation;
-  if (!CompareMatrices(total_wrench_in_world, momentum_rate_of_change, 1e-6,
-                       MatrixCompareType::absolute, &explanation)) {
+  if (!drake::CompareMatrices(total_wrench_in_world, momentum_rate_of_change,
+                              1e-6, drake::MatrixCompareType::absolute,
+                              &explanation)) {
     throw std::runtime_error("Drake:ValueCheck ERROR:" + explanation);
   }
 }
