+++ drake/systems/plants/test/compareParsersmex.cpp
@@ -6,12 +6,14 @@
 #include "drake/common/eigen_types.h"
 #include "drake/systems/plants/RigidBodyTree.h"
 #include "drake/util/drakeMexUtil.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 
 using namespace Eigen;
 using namespace std;
-using drake::util::MatrixCompareType;
+
+using drake::CompareMatrices;
+using drake::MatrixCompareType;
 
 /*
  * compares C++ robots generated via the matlab constructModelmex with the same
