+++ drake/systems/plants/test/testAccelerometer.cpp
@@ -3,7 +3,7 @@
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/common/drake_path.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 
 using Eigen::Vector3d;
@@ -13,7 +13,6 @@ using std::shared_ptr;
 using std::make_shared;
 using drake::GetDrakePath;
 using drake::RigidBodySystem;
-using drake::util::MatrixCompareType;
 using drake::RigidBodyAccelerometer;
 
 namespace drake {
