+++ drake/systems/plants/test/testGyroscope.cpp
@@ -2,7 +2,7 @@
 
 #include "drake/common/drake_path.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 
 using Eigen::Vector3d;
@@ -12,7 +12,6 @@ using std::shared_ptr;
 using std::make_shared;
 using drake::GetDrakePath;
 using drake::RigidBodySystem;
-using drake::util::MatrixCompareType;
 using drake::RigidBodyGyroscope;
 
 namespace drake {
