+++ drake/systems/plants/test/testIKpointwise.cpp
@@ -8,15 +8,14 @@
 #include "drake/systems/plants/IKoptions.h"
 #include "drake/systems/plants/RigidBodyIK.h"
 #include "drake/systems/plants/RigidBodyTree.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 
 using Eigen::MatrixXd;
 using Eigen::Vector2d;
 using Eigen::Vector3d;
 
-using drake::GetDrakePath;
-using drake::util::CompareMatrices;
-using drake::util::MatrixCompareType;
+namespace drake {
+namespace {
 
 GTEST_TEST(testIKpointwise, simpleIKpointwise) {
   RigidBodyTree model(
@@ -78,3 +77,6 @@ GTEST_TEST(testIKpointwise, simpleIKpointwise) {
   delete[] constraint_array;
   delete[] info;
 }
+
+}  // namespace
+}  // namespace drake
