+++ drake/systems/plants/test/testMagnetometer.cpp
@@ -3,7 +3,7 @@
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/common/drake_path.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 
 using Eigen::Vector3d;
@@ -14,7 +14,6 @@ using std::make_shared;
 using drake::GetDrakePath;
 using drake::RigidBodyMagnetometer;
 using drake::RigidBodySystem;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace systems {
