+++ drake/util/test/testDrakeGeometryUtil.cpp
@@ -6,7 +6,7 @@
 #include "drake/common/eigen_types.h"
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/util/drakeGeometryUtil.h"
-#include "drake/util/eigen_matrix_compare.h"
+#include "drake/common/eigen_matrix_compare.h"
 #include "drake/util/testUtil.h"
 #include "gtest/gtest.h"
 
@@ -20,7 +20,6 @@ using Eigen::Dynamic;
 using Eigen::Quaterniond;
 using Eigen::Translation3d;
 using std::default_random_engine;
-using drake::util::MatrixCompareType;
 
 namespace drake {
 namespace util {
